Remove trigram index on notes
This removes the GIN index on notes because these are expensive to
update. This causes note creation to sometimes time out.

This query is also seldom used. We only search for notes in the Global
Search page and only searches within projects are allowed. And when
searching within a project, this global index is unlikey to be used and
might even degrade performance as we see in issue searches. For issue
searches within a project, we use a CTE to avoid the global trigram
index.

Changelog: performance