Ensure selected projects are always visible
When displaying selected / unselected projects in the vsa
project selector, we were checking the `availableProjects`
computed property, this would return the items that match the
current search string, instead we should always use the
`selectedProjects` computed property for selected items.

This ensures that selected items that do not match the
current search term are still visible.

Adds a related spec test