Respect security scanner schema
As outlined in https://gitlab.com/gitlab-org/gitlab/-/issues/339507 our
GraphQL mutation deviated from the security scanner schema is some
places. This attempts to collect the same dataset but falls back to some
hardcoded values if the required ones are not provided. To do this we
have to:

* use externalId and externalType in Vulnerabilites::Identifier creation
* collect additional data for Vulnerabilites::Scanner creation

Changelog: changed
EE: true