Clean up migration to populate commit users
This cleans up any pending MigrateMergeRequestDiffCommitUsers migration
jobs. In addition, we stop using the old columns, ignore them, and
remove them in a post-deployment migration.

We don't need a separate release for the removal of the columns, as
ignoring them in the deploy ensures it's safe to remove the columns in a
post-deployment migration. In addition, the code has supported reading
from/writing to the new data for the last two or so months.

With this change we should start seeing less growth in the
merge_request_diff_commits table, as we no longer store many duplicate
names and Emails. Once we find a way to reclaim unused space in the
table, we should be able to free up roughly 500 GB of disk space as
mentioned in the original issue
(https://gitlab.com/gitlab-org/gitlab/-/issues/331823).

Changelog: added