Fix nil value in location column in vulnerability_occurrences
This change fixes the bug of persisting
nil value in location column of vulnerability_occurrences
table. Currently it does not break any other components
because if location is nil, it is fetched from raw_metadata
column. But in order to add index to location column, the
value has to be persisted.

EE: true
Changelog: fixed