test: Set Gitaly's bundle path as absolute path
Tests depend on Gitaly running in the background. There are a few
scripts and libraries that take care of setting up and spawning
Gitaly. These instructions run from different locations in the project
source tree.

Because some people might be specifying the BUNDLE_PATH as a relative
location (i.e. vendor/bundle), the setup step might install gems in a
different location as opposed to where Gitaly-ruby expects them in the
spawn step.

To overcome this, the BUNDLE_PATH is read from the Bundler configuration
and expanded to an absolute path before it's passed on to Gitaly
setup/spawn.