Use a YAML file to define gitlab_schema for tables
We were previously using the gitlab_schema annotation on models which
was causing too much issues with missed models and generally it was
trickier to get right due to autoloading issues among other fancy Ruby
metaprogramming trickery.

We also needed to add
`spec/migrations/disable_job_token_scope_when_unused_spec.rb` to the
allowlist as this had a previously undetected cross-join. The previous
code which used the model to determine `gitlab_schema` was incorrectly
detecting `gitlab_main` as the migration test declares the model and
overwrites the `gitlab_ci` value already defined.