test: Use GDK's Gitaly vendored gems if exist
Some users prefer to vendor the installed gems. This also applies to the
gems installed in <gdk>/gitaly/ruby.

Since [1] the setup of Gitaly is sped up by copying files over from
GDK's Gitaly. But this change also skipped over installation of gems. So
when GDK's Gitaly is using vendored gems, Gitaly in test will not find
them.

This change goes looking into GDK's Gitaly to see if it vendored gems,
and if so the BUNDLE_PATH is set to that directory.

1. 08af0f9f824 (Speed up gitaly setup in tests, 2021-07-20)