Ensure Azure-related gems are loaded
We attempted to load the gitlab-fog-azure-rm only if it was really
needed for object storage, but there are a number of cases where we
don't dynamically load these gems when we need them. To avoid the
complexity of this dynamic allocation, we now require the gem, just as
we do with other object storage providers. Unfortunately this adds 5 MB
to the RAM usage of the process according to derailed_benchmarks, but
this fairly typical for a gem for object storage.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/334853

Changelog: fixed