Fix `scripts/security-harness` to support older git versions
Prior this commit `scripts/security-harness` would not remove lefthook
if the git < 2.31 because the option `--path-format=absolute` wasn't
available yet and just been echoed which resulted in a multiline result:

  $ git rev-parse  --path-format=absolute --git-path hooks/pre-push
  --path-format=absolute
  .git/hooks/pre-push

The hook path "--path-format=absolute\n.git/hooks/pre-push" which does
not exist. Moreover, the first run creates a directory
"--path-format=absolute\n.git/" which is cumbersome to remove.

In this commit we only care about the last line which is most likely the
path to the hook.