Align the ordering of partitioned tables
In JiHu repository, there's an additional code to register this
table to be synchronized like:

    Gitlab::Database::Partitioning.register_models([
      Phone::VerificationCode
    ])

This does not present in GitLab repository, but only in JiHu repository,
and this will cause the table to move to the top of db/structure.sql along
with other synchronized partitioned tables.

This will align both repositories to be more consistent.

The same does also happen for FOSS where
incident_management_pending_alert_escalations and
incident_management_pending_issue_escalations are not aligned.