Don't verify module path for Naming/FileName cop
This causes false positives if the GitLab project is checked out inside
a directory called `src`, which Rubocop incorrectly detects as the
project root.

In practice this verification only seems to work correctly for certain
directories like `lib/` anyway, for `app/` it only considers the
filename and looks for a matching class definition, but ignores the
module namespace.

Also see https://github.com/rubocop/rubocop/issues/6289