Update error message of path conflict
Before transfering a group, it's checked if there is already Namespace
with the same path. Currently this effectively checks only if there is a
subgroup with the same path. If there is a project with the same path,
transfer fails too, but in the next step because updated group
validation fails with error `"Validation failed: Group URL has already
been taken"`.

When we create/backfill project namespace for each project, we will
effectively check both for subgroup or project path conflicts (because
project namespace will be created with the same path as project). So we
can just update the error message to mention that there is either group
or project with the same path.