Redirect Geo HTTP(s) pulls to different path with missing repo
Previously, we only redirected pushes to a different path on the
primary, however with unified URLs, redirecting pulls to the same
path will end in a redirect loop.

This changes the redirect so that pulls get also redirected to a
different path, that we always proxy to the primary.

Changelog: changed
EE: true