Use `GITLAB_LOAD_BALANCING_REUSE_PRIMARY_*` to allow primary re-use
The purpose of this feature is to allow Phase 3 of decomposition rollout
as described in https://gitlab.com/groups/gitlab-org/-/epics/6160#progress.

The idea is that we configure a separate DB (CI) with replicas, but due
to transaction visibility we need to share primary connection to `main`
database.

This adds a temporary feature enabled via `FF` to allow this type
of behavior change.