Fix 2+ discussions on one line not working
Always create a lines object, even if only one line is provided.

Some instances of creating the note form won't have a range of lines
at construction.

If only a single line is provided, create the range of lines using
that one line as both the start and end of the range.

Changelog: fixed