Fix `canUpdate` console error on external issues
When setting the `canUpdate` provided attribute
in issue_field.vue, we were referencing `this`,
which is undefined at the time when
we "provide" the data.

This commit updates Jira and ZenTao issue
trackers to have `canUpdate` set to false.

It also changes the behavior of issue_field. Now,
we pass a `canEdit` prop to determine when to
make a field editable or not.

Changelog: fixed
EE: true