Link Pipeline Editor button to open in correct branch
When previewing a gitlab config file in the repository view,
you can click on the "Pipeline Editor" button to open it in
the Pipeline editor section. If you were previewing the file
in a branch other than `main`, then it would not carry over
when landing in the section. This commit changes that.

Changelog: changed