Fix user mentions migration
The user mentions background migration uses an isolated copy of the
Group class.

We need to update this class to have the root_saml_provider method
because the policy check was changed to use this method instead of
root_ancestor.saml_provider directly. That was done to allow preloading
of SAML providers.