Remove the pipelines link from Auto DevOps toast
This removes the link to the pipelines page from the Auto DevOps toast.
Since our switch to BootsrapVue's toast component, HTML is escaped in
toasts' messages, meaning that this link could not render.
After discussing this issue, we have concluded that the link could be
removed altogether as toasts should probably not contain any interactive
elements for accessbility concerns.

Changelog: fixed