Revert temporary change for sending expiration email for all ssh keys
We started sending SSH key expiration notification emails for all
expired keys in the
MR https://gitlab.com/gitlab-org/gitlab/-/merge_requests/62114.
This change is no longer needed and, therefore, changing back to
sending notifications only for the keys that are expiring today.
Also removing the index that is not required anymore and
running the 'ssh_keys_expired_notification_worker' cron twice
daily as this is an idempotent job.

Changelog: changed