Standardize error message field in structured logs
Some of our log messages emitted `error` as a string, and others emitted
`error.message`. In Elasticsearch, if the `error.message` is mapping is
used, `error` is an object capable of holding other fields. As a result
of these mixed types, Elasticsearch would flag a mapping conflict and
dropped certain log messages, depending on which value arrived first for
the day.

To avoid this mapping conflict, we now standardize on `error.message`.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/344015

Changelog: fixed