Exclude internal API and Geo git_ssh routes from secondary proxying
The api/internal routes need to not be proxied to allow detection if
the repository exists locally for a project, and redirection to the
primary if not.

The geo/proxy_git_ssh route is used internally by gitlab-shell to
then proxy the git+ssh operations to the primary, using the internal
URL of the primary (hence why these don't get proxied and is intended
that they hit the secondary).

Changelog: changed
EE: true