Rename crm related policies and consider feature flag status
As part of the ongoing effort to introduce the customer relations
feature, we have come across a few challenges with the length of
the name. So we are starting to adopt the crm acronym. This MR
updates all of the policies to ensure the module is references.

In addition, we avoid the need to check the feature flag for
every crm interaction by including it in all of the policies.

Changelog: changed