Do now unescape branch name when deleting branch
Previously, when we unescape the branch name passed by frontend
(e.g. `test%2ffeature`), it can match a branch matching the
unescaped name (e.g. `test/feature`).

In a case wherein both `test%2ffeature` and `test/feature` branch
exists, the previous behavior will lead to a bug wherein
`test/feature` will be deleted if `test%2ffeature` was the one
being deleted.

The fix is to remove the unescape code so we're just finding for
the branch to delete without unescaping the branch name.

Changelog: fixed