Increase the deduplication TTL for future jobs
With this, the expiry for the deduplication key will be the time the
job was scheduled in the future + the configured deduplication time
for the job.

This makes sure that we do not expire before the job has even started,
which could happen for jobs scheduled in the future that aren't picked
up immediately after they should have been.

Changelog: fixed