Update scroll position to synchronize the Jump To Next button
When you scroll a discussion halfway up the page, the
Jump To Next Unresolved Discussion button will go to the
discussion after that one.

When you scroll the page such that a discussion scrolls down
through the middle of the page and below it, the Jump button
considers that discussion the "next" in line to be seen, and
will scroll to it.

Changelog: added