Enqueue new issues workers vs old ones
We've introduced new issues workers in a previous MR
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/72802
that now comply to NamespacedClass cop offences tracked in
issue https://gitlab.com/gitlab-org/gitlab/-/issues/321982.

Now that we have the new workers in production, we can change
enqueueing jobs with new workers.

We still keep the old workers around in case some of the jobs
are still in old queues and need to be processed by the old
workers. We'll be removing the old workers in a future release.