Fix error on environment rollback
While the rollback was still completed successfully, there was
previously an extra alert showing when rolling an environment back to a
previous deployment via the deployments page. This ensures that the
error is not seen, by stopping an error being thrown for missing the
retry path after the rollback occurs.

Changelog: fixed