Pluralize DevOps Report to DevOps Reports
DevOps Report (singular) actually refers to 2 features. That has caused
confusion. This changes it to be DevOps Reports (plural) in the UI, to
indicate it is a category. This aligns the app with the marketing
content, which already uses the plural version.

This also fixes some references to DevOps Report that should be
references to DevOps Adoption or DevOps Score.

Changelog: changed