Add default_branch_image to container scanning location
Adds a a new field named default_branch_image, which indicates the name
of the scanned image as it appears on the default branch. On non-default
branches, we will use this in lieu of the image field when creating the
location fingerprint. This will allow vulnerabilities to generate
identical fingerprints when the image name differs between the default
branch and the non-default branch.