Use database parameter to configure BG migrations
Introduce a new class to run background migration jobs, that properly
handles multiple databases by requiring the target trackign database
to be passed in.

This information will either be determined by metadata attached to DML
migrations, or from the appropriate worker class for the target
database. For now we default to :main to maintain
backwards-compatibility.