Default jira_use_first_ref_by_oid feature flag to true
In https://gitlab.com/gitlab-com/gl-infra/production/-/issues/5662, we
saw a lot of requests for `FindAllTags` that never completed. A project
with JIRA enabled would attempt to call:

```ruby
noteable.ref_names(project.repository).first
```

The `FindAllTags` triggered by `ProcessCommitWorker` comes from the JIRA
integration. This happens when the commit message references a JIRA
issue.

In https://gitlab.com/gitlab-org/gitaly/-/merge_requests/3947, we added
a new RPC `FindRefsByOID`. The purpose of this RPC is to, as the name
suggests, find refs by passing in an object id and some other optional
parameters.

This has been on production for several weeks now, and the RPC landed in
GitLab 14.5 so let's flip this feature on to true by default for 14.6.

Relates to:

* https://gitlab.com/gitlab-org/gitlab/-/issues/343585
* https://gitlab.com/gitlab-org/gitlab/-/issues/343035

Changelog: changed