Runner status won't return "paused"
This change removes a data duplication in the GraphQL API by removing
the `PAUSED` status as a possible result of the Ci::Runner `status`

The same change is announced for the REST API for "status", and "paused"
will be removed in the next major release.

In both cases, when checking if a runner is "paused", API users are
advised to check the boolean attribute "active" to be `false` instead.

Changelog: changed