Enable workhorse_use_sidechannel by default
This changes the workhorse_use_sidechannel feature flag to be on by
default. This is a performance improvement for Git fetch traffic, see
https://gitlab.com/groups/gitlab-com/gl-infra/-/epics/463.

For self-managed installations that have configured a gRPC proxy to
sit between Workhorse and Gitaly, this will break things. The solution
for that is to remove the proxy or change it to a TCP proxy. Having a
proxy between Workhorse and Gitaly is not a recommended configuration
but we cannot rule out that someone chose to do this. Also see
https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/1301.

This commit intentionally leaves the feature flag around so that in
case someone runs into this problem, they have a temporary workaround
(disabling the feature flag) besides the real solution (not using a
gRPC proxy).

Changelog: performance