Add License checks to AgentAuthorizationsFinder
The CI/CD tunnel's ci_user and ci_job impersonation features are
EE-only, so these should only be returned when the license is available.
We check the license on the CI job's project under the assumption that
this project and the agent fall under the same license.

Part of https://gitlab.com/gitlab-org/gitlab/-/issues/327848