Change trailing slash from constraints to default for insights routes
Simply specifying trailing_slash: true as a parameter to the route
matcher adds it as both a constraint and option.

However this means url_for will not be able to generate this dynamically
since trailing_slash is a reserved keyword for url_for and cannot be
used as an option.

This makes the _default_, not constraint to have a trailing slash, so
that path and URL helpers generate it _with_ the trailing slash, but
not _require_ it.