Allow percentage rollout of load balancer of decomposed database
This introduces a `GITLAB_USE_MODEL_LOAD_BALANCING`
and `use_model_load_balancing` to enable percentage rollout
of used connection when accessing data.

The usage pattern is to redirect 1% of new connections using CI DB
to use a Load Balancer replicas (and presumbly later primary connection).

This is achieved by injecting a FF evaluation logic as part of LB setup
to re-define how connection is fetched.