test: Expand bundle path to absolute path
In previous commit we started setting BUNDLE_PATH as an environment
variable, and this works great. But this value gets overridden in case
it's specified in the local config. You can see so by using the
`bundle config` command:

    $ BUNDLE_PATH=vendor/bundle bundle config path
    Settings for `path` in order of priority. The top value will be used
    Set for your local app (/home/r/gdk/gitlab/.bundle/config): "vendor/bundle"
    Set via BUNDLE_PATH: "vendor/bundle"
    Set for the current user (/home/r/.config/bundle): "vendor/bundle"

This isn't persay an issue, but in most cases this is set as a relative
path. Because different `bundle` commands might get triggered from
different directory levels in the project tree, this might become an
issue.

To overcome this, the active value of BUNDLE_PATH is read, expanded to
an absolute path and passed back to the environment. But because of the
prioritization demonstrated above, we need to set BUNDLE_IGNORE_CONFIG
so any local, user, or system configuration is ignored.