Fix quick action permissions to match UI
This changes permissions on some quick actions to match what we check in
the issuable base service.

We recently changed the check in the service so that guests can set
labels, milestones, and other metadata when creating an issue. This
allows them to set those via quick actions.

Changelog: fixed