Toggle file tree bolding based on file "Viewed" status
Before, when scrolling to a file in the MR Review OR when
clicking the file title in the file tree view, the file would
be marked as "viewed" (and thus unbolded).

Now, the only action that can signify that a file is viewed
(and unbold the entry in the tree) is clicking the "Viewed"
checkbox on the file header in the Diffs view.

Changelog: changed