Reorganise Jira Connect UI
The commit reorganises the UI logic for the Jira Connect
App. This is part-UI enhancement, part tech-debt cleanup.

To the user, the main change is that the Add Namespace button
is now in the empty state when the user is signed in but has no
subscriptions.

This restructuring also enables us to move some HAML
markup to Vue.

Changelog: changed