Wrap Sidekiq scheduler threads in Rails reloader
We execute DB queries in our Sidekiq client middleware so this is needed
to properly checkin DB connections after every poll. On the next poll, a
new connection will be checked out and a connection check query will be
executed. Rails would then handle the reconnection for us in case the
connection is bad.

Without this, we would continue using a bad connection until Sidekiq is
restarted.

Changelog: fixed