spec: Do not recreate custom repositories
There are two project factories which create repositories, once with a
default repository and once with custom files which then get written
into the repository. These can be used in combination to use a default
repository which has a bunch of custom files on top, and some tests do
in fact do this. When doing so, we call Gitaly's `CreateRepository()`
RPC twice though, where the second call is idempotent and doesn't really
do anything interesting.

Gitaly is about to change behaviour of `CreateRepository()` to fail in
case it is asked to recreate an already existing repository. So let's
call `CreateRepository()` for custom repos only when they do not exist
yet.