spec: Do not recreate existing snippet repositories
In order to set up snippet repositories in some of our tests, we first
copy over a template repository into the target location and then call
Gitaly's `CreateRepository()`. While this call wouldn't do anything if
used with Gitaly only given that the RPC is idempotent, it would create
a database record when Gitaly is proxied by Praefect.

Gitaly is about to drop idempotentcy of `CreateRepository()`, where it
will instead raise an error. Prepare for this by reversing calls such
that we first create the repository in Gitaly, and then copy over
contents of our seed repository.