Prevent Git operations from checking replication lag on non-Geo sites
We were checking if the database is read only, and verifying the
replication lag, however with the addition of maintenance mode, the
read only check is also true when maintenance mode is enabled.

The problem with only using the read only check is two-fold:
unnecessary replication lag queries, when the current site is not
a secondary, as maintenance mode doesn't pause any DB operations;
and breaking databases that don't support the standard replication
functions to check the lag (i.e. Aurora).

Changelog: fixed
EE: true